refactor(workboard): split board filter helpers
