+++ extensions/workboard/src/gateway.ts
@@ -1,9 +1,10 @@
-import { WORKBOARD_STATUSES, type WorkboardCard } from "@openclaw/workboard-contract";
+import type { WorkboardCard } from "@openclaw/workboard-contract";
 // Workboard plugin module implements gateway behavior.
 import type { OpenClawPluginApi } from "../api.js";
 import {
   assertNoCursorAdvance,
   createWorkboardDispatchHandler,
+  listWorkboardCards,
   readId,
   respondError,
 } from "./gateway-helpers.js";
@@ -58,15 +59,7 @@ export function registerWorkboardGatewayMethods(params: {
     "workboard.cards.list",
     async ({ params: requestParams, respond }) => {
       try {
-        const [cards, { boards }] = await Promise.all([
-          store.list({ boardId: requestParams.boardId }),
-          store.listBoards(),
-        ]);
-        respond(true, {
-          cards: cards.map(redactClaimToken),
-          boards,
-          statuses: WORKBOARD_STATUSES,
-        });
+        respond(true, await listWorkboardCards(store, requestParams.boardId, redactClaimToken));
       } catch (error) {
         respondError(respond, error);
       }
