+++ ui/src/pages/workboard/view.ts
@@ -7,7 +7,6 @@ import type { AgentsListResult, GatewaySessionRow } from "../../api/types.ts";
 import { icons } from "../../components/icons.ts";
 import "../../components/modal-dialog.ts";
 import "../../components/tooltip.ts";
-import "../../components/web-awesome-select.ts";
 import { t } from "../../i18n/index.ts";
 import { formatDateMs, formatDateTimeMs, formatDurationCompact } from "../../lib/format.ts";
 import "../../styles/workboard.css";
@@ -61,13 +60,7 @@ import {
   normalizeActiveBoardFilter,
   WORKBOARD_ALL_BOARDS_FILTER,
 } from "./board-filter.ts";
-
-type WorkboardSelectOption<Value extends string = string> = {
-  value: Value;
-  label: string;
-  description?: string;
-  disabled?: boolean;
-};
+import { renderWorkboardSelect, type WorkboardSelectOption } from "./workboard-select.ts";
 
 type WorkboardProps = {
   host: object;
@@ -447,66 +440,6 @@ function isCardActionTarget(event: Event): boolean {
   return event.target instanceof Element
     ? Boolean(event.target.closest("button, a, input, select, textarea"))
     : false;
-}
-
-function renderWorkboardSelect<Value extends string>(params: {
-  value: Value;
-  options: readonly WorkboardSelectOption<Value>[];
-  label: string;
-  onChange: (value: Value) => void;
-  requestUpdate?: () => void;
-  className?: string;
-  showLabel?: boolean;
-  disabled?: boolean;
-}) {
-  const select = html`
-    <wa-select
-      class="workboard-select ${params.className ?? ""}"
-      label=${params.label}
-      value=${params.value}
-      ?disabled=${params.disabled}
-      @change=${(event: Event) => {
-        const value = (event.currentTarget as HTMLElement & { value?: string }).value as
-          | Value
-          | undefined;
-        if (
-          value !== undefined &&
-          params.options.some((option) => option.value === value && !option.disabled)
-        ) {
-          params.onChange(value);
-          params.requestUpdate?.();
-        }
-      }}
-    >
-      ${params.options.map(
-        (option) => html`
-          <wa-option
-            class="workboard-select__option"
-            value=${option.value}
-            .label=${option.label}
-            ?selected=${option.value === params.value}
-            ?disabled=${option.disabled}
-          >
-            <span class="workboard-select__copy">
-              <span class="workboard-select__label">${option.label}</span>
-              ${option.description
-                ? html`<span class="workboard-select__description">${option.description}</span>`
-                : nothing}
-            </span>
-          </wa-option>
-        `,
-      )}
-    </wa-select>
-  `;
-  if (params.showLabel === false) {
-    return select;
-  }
-  return html`
-    <div class="workboard-field">
-      <span>${params.label}</span>
-      ${select}
-    </div>
-  `;
 }
 
 function engineDisplayName(engine: WorkboardExecutionEngine): string {
